change paging to Long

+++ blibli-backend-framework-common/src/main/java/com/blibli/oss/backend/common/helper/PagingHelper.java
@@ -5,15 +5,15 @@ import com.blibli.oss.backend.common.model.response.Paging;
 
 public class PagingHelper {
 
-  public static Paging toPaging(PagingRequest request, Integer totalItem) {
-    int totalPage = totalItem / request.getItemPerPage();
+  public static Paging toPaging(PagingRequest request, Long totalItem) {
+    Long totalPage = totalItem / request.getItemPerPage();
     if (totalItem % request.getItemPerPage() != 0) {
       totalPage++;
     }
     return toPaging(request, totalPage, totalItem);
   }
 
-  public static Paging toPaging(PagingRequest request, Integer totalPage, Integer totalItem) {
+  public static Paging toPaging(PagingRequest request, Long totalPage, Long totalItem) {
     return Paging.builder()
       .page(request.getPage())
       .totalItem(totalItem)
